+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+* Support for Region ID from regions endpoint
+
+
 ## 0.41.0 (3rd November 2025)
 
 ### Added: