+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/).
 
-## Unreleased
+## 0.42.0 (10th November 2025)
 
 * Added `ID` field to `Region` struct in account model
 * Added Redis version support for Essentials databases: `RedisVersion` field and `UpgradeRedisVersion()` method